+++ acwandroid/app/src/main/java/com/zapp/acw/bll/Card.java
@@ -3,9 +3,9 @@ package com.zapp.acw.bll;
 import java.util.ArrayList;
 
 public final class Card {
-	public int cardID = 0;
-	public int noteID = 0;
-	public int modelID = 0;
+	public long cardID = 0;
+	public long noteID = 0;
+	public long modelID = 0;
 	public ArrayList<String> fieldValues = new ArrayList<> ();
 	public String solutionFieldValue;
 }
