+++ acwandroid/app/src/main/java/com/zapp/acw/bll/Field.java
@@ -1,6 +1,6 @@
 package com.zapp.acw.bll;
 
 public final class Field {
-	public int idx = 0;
+	public long idx = 0;
 	public String name;
 }
